requirements.txt
aiohttp==3.8.1
discord.py==2.0.0
spacy==3.2.1
en-core-web-lg==3.1.0
sqlite3==2.6.0
python-dotenv==0.19.2
openai==0.27.0
numpy==1.21.5
scikit-learn==1.0.1
